Update to parseInt

+++ statistics-dashboard-worker/src/main/java/eu/europeana/statistics/dashboard/worker/execution/TargetDataRunner.java
@@ -72,10 +72,10 @@ public final class TargetDataRunner implements CommandLineRunner {
     for (int i = 1; i < targetData.size(); i++) {
       List<String> row = targetData.get(i);
       LOGGER.info("Started writing data of country {} into database", row.get(0));
-      Target firstModel = new Target(row.get(0), Long.parseLong(row.get(3)),
-          Long.parseLong(row.get(2)), Long.parseLong(row.get(1)), 2025);
-      Target secondModel = new Target(row.get(0), Long.parseLong(row.get(6)),
-          Long.parseLong(row.get(5)), Long.parseLong(row.get(4)), 2030);
+      Target firstModel = new Target(row.get(0), Integer.parseInt(row.get(3)),
+          Integer.parseInt(row.get(2)), Integer.parseInt(row.get(1)), 2025);
+      Target secondModel = new Target(row.get(0), Integer.parseInt(row.get(6)),
+          Integer.parseInt(row.get(5)), Integer.parseInt(row.get(4)), 2030);
       results.add(firstModel);
       results.add(secondModel);
     }
